[FRD-39] Change component architecture folders
## [FRD-39] Description
Moved Card, Header, and related files into common and headers folders for better organization. Updated all relevant imports and exports to use new paths and created index files for easier imports. Adjusted test imports and SCSS usage to match the new structure. Added line-height to global styles.
