+++ src/components/Chat/ChatInput/ChatInput.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render, screen, fireEvent } from '@testing-library/react';
-import ChatInput from '.';
+import ChatInput from './ChatInput';
 import { Provider } from 'react-redux';
 import configureStore from 'redux-mock-store';
