+++ src/components/buttons/index.jsx
@@ -1,0 +1,1 @@
+export { default as CTAButton } from './CTAButton/CTAButton';
