+++ src/components/common/Card/Card.scss
@@ -1,4 +1,4 @@
-@use '../../style/variables' as *;
+@use '../../../style/variables' as *;
 
 .card {
    background: $white;
