+++ src/components/common/Card/Card.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render, screen } from '@testing-library/react';
-import Card from '.';
+import Card from './Card';
 
 describe('Card Component', () => {
    it('Renders children', () => {
