+++ src/components/common/Card/Card.types.ts
