+++ src/components/common/index.tsx
@@ -1,0 +1,1 @@
+export { default as Card } from './Card/Card';
