+++ src/style/globals.scss
@@ -9,6 +9,7 @@ body {
 body {
   font-family: Arial, Helvetica, sans-serif;
   font-size: 14px;
+  line-height: 1.3rem;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
   background: $page-background;
